requirements.txt
statsmodels
streamlit
pandas
numpy
matplotlib
seaborn
yfinance
scikit-learn
tensorflow-cpu==2.18.0
joblib